requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
psycopg==3.2.12
drf-yasg==1.21.11
gunicorn==23.0.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
psycopg2-binary
django-templated-email
requests
dj-database-url
whitenoise
python-decouple
django-cors-headers


#linux
#sudo apt update
#sudo apt install postgresql
#sudo apt install postgresql-contrib
